refactor: change files location
